Add temporary conversation data
- Import the `Conversation` type from the "ui" module.
- Create a temporary array variable `temp` containing a single
conversation object.
- In the WebSocket `open` function, send the first conversation from
`temp`.

refactor: Update API communication

- Import the `Message` type from the "ui" module.
- Replace the default export in "api.ts" with `chat` const and
`sendMessage` function.
- Create the `_messages` writable store and initialize it as an empty
array.
- Subscribe to the chat and add received messages to the store.
- Implement the `sendMessage` function to add a new message to the
store.
