+++ apps/wave-backend/src/index.ts
@@ -1,4 +1,7 @@
 import { Elysia, t } from "elysia";
+import type { Conversation } from "ui";
+
+let temp: Conversation[] = [{ name: "Convo 1", id: 123456 }];
 
 const app = new Elysia()
     .get("/", () => "Hello Elysia")
@@ -10,6 +13,9 @@ const app = new Elysia()
         }),
         open(ws) {
             const msg = `${ws.remoteAddress} has entered the chat`;
+
+            ws.send(temp[0]);
+
             ws.subscribe("the-group-chat");
             ws.publish("the-group-chat", { message: msg });
         },
